TECH-4526: move command to hobo_fields/extensions
