+++ lib/hobo_fields/extensions/module.rb
